wayd_neuron_validator/requirements.txt
keras==2.8.0
numpy==1.22.3
Flask==2.0.3
opencv-python==4.5.5.64
kafka-python==2.0.2
tensorflow==2.8.0
python-dotenv==0.19.2